backend/requirements.txt
django>=5.0.0
djangorestframework
django-cors-headers
psycopg2-binary
dj-database-url
whitenoise
gunicorn
python-dotenv
Pillow